+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Era.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE CPP #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 #if __GLASGOW_HASKELL__ >= 908
@@ -13,8 +14,10 @@ module Cardano.Ledger.Alonzo.Era (
   AlonzoUTXOW,
   AlonzoBBODY,
   AlonzoLEDGER,
+  hardforkConwayTranslateUpperBoundForPlutusScripts,
 ) where
 
+import Cardano.Ledger.BaseTypes (ProtVer (pvMajor), natVersion)
 import Cardano.Ledger.Internal.Era (AlonzoEra)
 import Cardano.Ledger.Mary (MaryEra, MaryValue)
 import Cardano.Ledger.Shelley.Core
@@ -90,3 +93,8 @@ type instance EraRule "TICK" AlonzoEra = ShelleyTICK AlonzoEra
 type instance EraRule "TICKF" AlonzoEra = ShelleyTICKF AlonzoEra
 
 type instance EraRule "UPEC" AlonzoEra = ShelleyUPEC AlonzoEra
+
+-- | Starting with protocol version 9, we translate the upper bound of validity
+-- interval correctly for Plutus scripts.
+hardforkConwayTranslateUpperBoundForPlutusScripts :: ProtVer -> Bool
+hardforkConwayTranslateUpperBoundForPlutusScripts pv = pvMajor pv > natVersion @8